Rename measure.py to measurement.py
