+++ mindtrace/hardware/mindtrace/hardware/cameras/homography/__init__.py
@@ -1,5 +1,5 @@
 from mindtrace.hardware.cameras.homography.calibration import HomographyCalibrator, CalibrationData
-from mindtrace.hardware.cameras.homography.measure import PlanarHomographyMeasurer, MeasuredBox
+from mindtrace.hardware.cameras.homography.measurement import PlanarHomographyMeasurer, MeasuredBox
 
 __all__ = [
     "HomographyCalibrator",
